added new structure
